Biaschecker/news-bias-classifier/requirements.txt
transformers
torch
gradio
newspaper3k